% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{filter_by_style}
\alias{filter_by_style}
\title{Sort/filter transport options according to the user's travel style.}
\usage{
filter_by_style(options, style)
}
\arguments{
\item{options}{A list of `transport_option` objects.}

\item{style}{One of `"fastest"`, `"cheapest"`, `"scenic"`.}
}
\value{
The list, sorted ascending by the relevant key (and trimmed).
}
\description{
Sort/filter transport options according to the user's travel style.
}
