% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripPlanner-package.R
\docType{package}
\name{tripPlanner-package}
\alias{tripPlanner}
\alias{tripPlanner-package}
\title{tripPlanner: Optimal trip routing and transport suggestions.}
\description{
The Nijat module of the multi-city trip planner. It provides:
\itemize{
  \item a fast Travelling Salesman solver implemented in C++ via Rcpp,
  \item an [R6::R6Class()] [RouteOptimizer] orchestrating the trip,
  \item transport-leg suggestions filtered by user preferences.
}
}
\author{
\strong{Maintainer}: Kamilla Tadjibaeva \email{kamilla.tadjibaeva2020@gmail.com}

Authors:
\itemize{
  \item Nijat Kazimli \email{nctkzmli@gmail.com}
}

}
\keyword{internal}
